Implement claims

+++ resources/store.py
@@ -1,6 +1,6 @@
 ## Import libraries
 from flask_restful import Resource
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.store import StoreModel
 from typing import Dict
 
@@ -37,24 +37,30 @@ class Store(Resource):
         return store.json(), 201
 
     @jwt_required
-    def delete(self, name) -> Dict:
+    def delete(self, name) -> tuple:
         """
 
         :param name:
         :return:
         """
+        ## Check is_admin
+        claims = get_jwt_claims()
+
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege is required'}, 404
+
         store = StoreModel.find_by_name(name)
         if store:
             store.delete_from_db()
 
-        return {'message': 'Store deleted'}
+        return {'message': 'Store deleted'}, 200
 
 ## StoreList Class
 class StoreList(Resource):
     @jwt_required
-    def get(self) -> Dict:
+    def get(self) -> tuple:
         """
 
         :return:
         """
-        return {'stores': [store.json() for store in StoreModel.find_all()]}
+        return {'stores': [store.json() for store in StoreModel.find_all()]}, 200
